Fix Swift 6.2.3 type-checker failures under Xcode 26.3
These are Swift 6.2.3 (Xcode 26.3) inference failures, not deployment-target
issues — they reproduce at 26.4, 26.0, and 18.0. Xcode 26.4 / Swift 6.3
resolves them, which is why upstream does not see them. Cherry-pickable as-is.

+++ wisp/SwipeLeftToPop.swift
@@ -56,7 +56,7 @@ struct SwipeBackFromLeftEdgeModifier: ViewModifier {
                             .offset(x: -((geo.size.width - dragX) / 3.0))
                             .overlay(
                                 Color.black.opacity(
-                                    0.25 * (1 - min(1, dragX / geo.size.width))
+                                    0.25 * (1 - min(1, Double(dragX / geo.size.width)))
                                 )
                             )
                     }
